renaming GithubRepositoryCloneFlow
