+++ editor/src/components/editor/editor-component.tsx
@@ -71,7 +71,7 @@ import { EditorModes } from './editor-modes'
 import { allowedToEditProject } from './store/collaborative-editing'
 import { useDataThemeAttributeOnBody } from '../../core/commenting/comment-hooks'
 import { CollaborationStateUpdater } from './store/collaboration-state'
-import { GithubRepositoryCloneFlow } from '../github/github-clone-overlay'
+import { GithubRepositoryCloneFlow } from '../github/github-repository-clone-flow'
 
 const liveModeToastId = 'play-mode-toast'
